--- Mylist.rb
@@ -1,25 +1,0 @@
-require_relative 'MyEnumerable'
-
-class MyList
-  include MyEnumerable
-  def initialize(*list)
-    @list = list
-  end
-
-  def each(&block)
-   @list.each(&block) 
-  end
-end
-
-list = MyList.new(1, 2, 3, 4)
-
-puts list.all? {|e| e < 5}
-# => true
-puts list.all? {|e| e > 5}
-# => false
-puts list.any? {|e| e == 2}
-# => true
-puts list.any? {|e| e == 5}
-# => false
-puts list.filter {|e| e.even?}
-# => [2, 4]
